requirements.txt
flask
pydantic
pytest
pika
mysql-connector-python